EverydayWechat/requirements.txt
requests
beautifulsoup4
itchat
apscheduler
pyyaml
